SystemdPlugin must be explicitly enabled
Based on what others have done, this seems to be a change requierd when upgrading the version of sbt-native-packager used by Play 2.5 vs Play 2.6 (v1.1 to v1.3), but I haven't found release notes to confirm this.

Without this change, the systemd unit file isn't created so the service never starts.