Upgrade to Panda v5.0.0
Although this release is a major-version-bump (from 4.0.0 to 5.0.0) this new
version has no new functionality, just a few refactors and some dependency changes.
The `gha-scala-library-release-workflow` detected potential binary-incompatible
changes, and so allocated a major-version-bump, but any project upgrading to
Panda v5 from v4 will not need to make any code changes to adapt to the new version.

https://github.com/guardian/pan-domain-authentication/releases/tag/v5.0.0